calendar_integration/requirements.txt
Flask>=3.0.0
requests>=2.32.3
python-dotenv>=1.0.1
openai>=1.32.0
asgiref>=3.7.0
structlog>=23.1.0
schedule>=1.2.0
Flask-WTF>=1.2.1
# Calendar Integration Dependencies 
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
# Weather Integration Dependencies
requests>=2.32.3 